Added integration tests
Added ArgentPony.WarcraftClient.Integration.Tests.csproj to the solution.  This contains tests that make live calls to the APIs.  It also uses custom attributes to enable retries when the Blizzard APIs throttles requests.  Also updated the types of properties in a few of the model classes based on errors discovered while running the integration tests.

+++ src/ArgentPonyWarcraftClient/Models/ProfileApi/CharacterStatistics/RatingStatisticWithoutValue.cs
@@ -17,6 +17,6 @@ namespace ArgentPonyWarcraftClient
         /// Gets the rating bonus.
         /// </summary>
         [JsonPropertyName("rating_bonus")]
-        public int RatingBonus { get; set; }
+        public double RatingBonus { get; set; }
     }
 }
